build: enable -Wall -Wextra as a CI gate and fix what it reports
Nothing in the build or in CI enabled -Wall: AM_CFLAGS inherits only what
apxs supplies (-Werror=implicit-function-declaration, -Wformat,
-Werror=format-security) and the CI configure added no CFLAGS.

Enable -Wall -Wextra -Werror in the build job (gcc) and in static-analysis
(clang, which keeps the two clang-only type checks on top), rather than in
the shipped AM_CFLAGS so distro package builds keep their own flags. Two
categories are excluded: -Wunused-parameter, since the Apache hook/callback
and config accessor signatures are fixed by their function-pointer types
(the same reason the SonarCloud S1172 findings are accepted), and
-Wmissing-field-initializers, since auth_openidc_module deliberately leaves
httpd's trailing `flags` member implicit - httpd only grew it during 2.4,
so spelling it out would break the older httpd in the distro matrix.

Fixed in src:
- signed/unsigned comparisons against the negative OIDC_CONFIG_POS_INT_UNSET
  sentinel in the generated provider getters, and against the int-valued
  option table in cfg/dir.c (an enum with only non-negative enumerators has
  an unsigned underlying type). Both sides already round-tripped through the
  same conversion, so behaviour is unchanged.
- the two authz handler tables are typed oidc_json_type_t; the separator
  table gets a struct of its own with a char field, instead of reusing the
  JSON-handler struct to hold a character - which is what the two "there's
  some overloading going on here" comments were apologising for.
- five -Wmaybe-uninitialized locals passed by address to a parser that only
  writes them on success: seeded with their documented defaults. These only
  appear in an optimized gcc build (reported from LTO), not in a syntax-only
  check.

Fixed in test (the gate builds these too):
- test_jose.c declared five `const char[512]` buffers and wrote them through
  snprintf((char *)buf, ...), i.e. modified a const object; the bogus const
  and the casts are gone.
- seven 64->32 truncations in stub.c and test-cmd.c made explicit.

+++ test/test-cmd.c
@@ -299,7 +299,7 @@ int enckey(int argc, char **argv, apr_pool_t *pool) {
 	}
 
 	cjose_err cjose_err;
-	int src_len = cjose_jwk_get_keysize(jwk->cjose_jwk, &cjose_err) / 8;
+	int src_len = (int)(cjose_jwk_get_keysize(jwk->cjose_jwk, &cjose_err) / 8);
 	int enc_len = apr_base64_encode_len(src_len);
 	char *b64 = apr_palloc(r->pool, enc_len);
 	apr_base64_encode(b64, (const char *)cjose_jwk_get_keydata(jwk->cjose_jwk, &cjose_err), src_len);
@@ -329,8 +329,8 @@ int hash_base64url(int argc, char **argv, apr_pool_t *pool) {
 			return -1;
 		}
 		oidc_jose_error_t err;
-		if (oidc_jose_hash_and_base64url_encode(r->pool, algo, (const char *)bytes, outlen, &output, &err) ==
-		    FALSE) {
+		if (oidc_jose_hash_and_base64url_encode(r->pool, algo, (const char *)bytes, (int)outlen, &output,
+							&err) == FALSE) {
 			fprintf(stderr, "oidc_jose_hash_and_base64url_encode failed: %s", err.text);
 			return -1;
 		}
@@ -372,7 +372,7 @@ int uuid(int argc, char **argv, apr_pool_t *pool) {
 	oidc_session_t z;
 
 	if (argc > 2) {
-		n = _oidc_str_to_int(argv[2], n);
+		n = (unsigned long)_oidc_str_to_int(argv[2], (int)n);
 		if (n > 25000000 * 10)
 			n = 25000000;
 	}
